# server/openapi.yaml
openapi: 3.1.0
info:
  title: NAAFTrack API
  version: 1.0.0
  description: API métier pour la surveillance forestière et les opérations terrain.
servers:
  - url: http://localhost:8787
paths:
...
  /api/ai/analyze:
    post:
      operationId: analyzeRisk
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - prompt
              properties:
                prompt:
                  type: string
                  minLength: 10
                provider:
                  type: string
                  enum:
                    - auto
                    - local
                    - openai
                    - gemini
                  default: auto
      responses:
        "200":
          description: Analyse locale ou fournie par une IA configurée
        "400":
          description: Requête invalide
  /api/incidents:
    post:
      operationId: createIncident
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - description
                - severity
                - latitude
                - longitude
              properties:
                title:
                  type: string
                description:
                  type: string
                type:
                  type: string
                  enum:
                    - wildfire
                    - deforestation
                    - degradation
                severity:
                  type: string
                  enum:
                    - low
                    - medium
                    - high
                    - critical
                latitude:
                  type: number
                longitude:
                  type: number
                areaHectares:
                  type: number
                  minimum: 0
      responses:
        "201":
          description: Incident créé
        "400":
          description: Données invalides
...